[CONJ-950] correct ResultSetMetaData.getPrecision() and getColumnType() for TINYTEXT/TEXT/MEDIUMTEXT/LONGTEXT
precision correction :
* was returning the byte length, not character length.
* when unknown/too long will return 0, not -1

column type :
* the columnType value for LONGBLOB/LONGTEXT will return correct Type value. (example LONGTEXT data will have Types.LONGVARCHAR, not Types.VARCHAR).

+++ src/main/java/org/mariadb/jdbc/client/Column.java
@@ -123,7 +123,7 @@ public interface Column {
    *
    * @return precision
    */
-  long getPrecision();
+  int getPrecision();
 
   /**
    * return column type from column server type
